+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 
 ### Added
+    
+January 2026
+
+- Added `radioFolder` list style option for KML/KMZ export in "multiple events" time mode.
 
 December 2025